+++ earc/cart_app/views.py
@@ -16,12 +16,14 @@ class cart_management:
         if not request.user.is_authenticated or request.user.is_active is False: 
             return redirect('user_app:user_login')
         
-        owner_obj = Owner.objects.get(customer=request.user, stage=0)
+
+        owner_obj = Owner.objects.get(customer=request.user)
+        
         
         if owner_obj:
-            cart_items = Cart.objects.filter(cart_customer=owner_obj)   
+            cart_items = Order.objects.filter(order_customer = owner_obj, status=0)   
         else:
-            cart_items = Cart.objects.none()
+            cart_items = Order.objects.none()
 
 
         context = {
@@ -33,6 +35,7 @@ class cart_management:
     def add_cart(request):
         if not request.user.is_authenticated or request.user.is_active is False: 
             return redirect('user_app:user_login')
+        
         if request.method == 'POST':
             color_id = request.POST.get('color_id')
             storage_id = request.POST.get('storage_id')
@@ -45,29 +48,35 @@ class cart_management:
             try:
                 owner_obj, created = Owner.objects.get_or_create(
                     customer = request.user,
-                    stage = 0
                 )
 
-                cart_obj, cart_created = Cart.objects.get_or_create(
-                    cart_customer = owner_obj,
+                cart_obj, cart_created = Order.objects.get_or_create(
+                    order_customer = owner_obj,
                     product = product_obj,
                     color = color_obj, 
-                    storage = storage_obj
+                    storage = storage_obj,
+                    status = 0
                 )
 
                 if cart_created:
-                    cart_obj.quantity = quantity
-                    cart_obj.save()
+                    if cart_obj.storage.stock < quantity:
+                        cart_obj.quantity = cart_obj.storage.stock
+                        cart_obj.save()
+                    else:
+                        cart_obj.quantity = quantity
+                        cart_obj.save()
                 else:
-                    cart_obj.quantity = cart_obj.quantity + quantity
-                    cart_obj.save()
+                    if cart_obj.storage.stock < (cart_obj.quantity + quantity):
+                        cart_obj.quantity = cart_obj.storage.stock
+                        cart_obj.save()
+                    else:
+                        cart_obj.quantity = cart_obj.quantity + quantity
 
                 context = {
                     'status':True,
                 }
                 return JsonResponse(context, safe=True)
             except Exception as e:
-                print('This is the error', e)
                 context = {
                     'status':False
                 }
@@ -80,7 +89,7 @@ class cart_management:
             return redirect('user_app:user_login')
 
         try:
-            cart_obj = Cart.objects.get(cart_id=id)
+            cart_obj = Order.objects.get(cart_id=id)
             if cart_obj:
                 cart_obj.delete()
                 context = {
@@ -105,12 +114,15 @@ class cart_management:
 
 
     def update_quantity(request):
+        if not request.user.is_authenticated or request.user.is_active is False: 
+            return redirect('user_app:user_login')
+        
         if request.method == 'POST':
             data = json.loads(request.body)
             qty = data.get('qty')
             id = data.get('cart_id')
             try:
-                cart_obj = get_object_or_404(Cart, cart_id=id)
+                cart_obj = get_object_or_404(Order, cart_id=id)
                 if cart_obj:
                     if cart_obj.storage.stock < qty:
                         context = {
@@ -124,12 +136,13 @@ class cart_management:
                         cart_obj.save()
                         product_price = cart_obj.product.price - cart_obj.product.discount_price
                         product_price = product_price + int(cart_obj.storage.price_of_size)
-                        toatl = product_price * cart_obj.quantity
+                        total = product_price * cart_obj.quantity
+                        cart_obj.total_price = total
                         
                         context = {
                             'status':True,
                             'stock': cart_obj.storage.stock,
-                            'total': toatl
+                            'total': total
                         }
                         return JsonResponse(context, safe=True)
                 else:
@@ -146,9 +159,12 @@ class cart_management:
                 return JsonResponse(context, safe=True)
 
     def clear_cart(request):
+        if not request.user.is_authenticated or request.user.is_active is False: 
+            return redirect('user_app:user_login')
+        
         try:
-            owner_obj = Owner.objects.get(customer=request.user, stage=0)
-            cart_obj = Cart.objects.filter(cart_customer=owner_obj)
+            owner_obj = Owner.objects.get(customer=request.user)
+            cart_obj = Order.objects.filter(order_customer=owner_obj, status=0)
             if cart_obj:
                 cart_obj.delete()
                 context = {
@@ -174,8 +190,8 @@ class cart_management:
     def update_subtotal(request):
             total = 0
             try:
-                owner_obj = Owner.objects.get(customer=request.user, stage=0)
-                cart_obj = Cart.objects.filter(cart_customer=owner_obj)
+                owner_obj = Owner.objects.get(customer=request.user)
+                cart_obj = Order.objects.filter(order_customer=owner_obj, status=0)
                 for item in cart_obj:
                     product_total = item.product.price - item.product.discount_price
                     product_total = product_total + int(item.storage.price_of_size)
